Relevant: db_echse_input: Deleted all SVC and minor fix
Adjusted db_echse_input(): Deleted all SVC relevant information as this is not needed anymore within the new ECHSE-WASA structure and related parameters to horizon specific parameters. Furthermore, a small bug was fixed when calculating cumulated depth of soil horizons (suddenly an error appeared which was not present before, don't know the reason).